Fix a TOCTOU.
When partial content(with Range requests in HTTP, REST cmd in FTP)
support was added, a stat(2) check for offset was done before open(2)
to determine if O_APPEND needs to be included with the open(2)
flags. stat(2) before an open(2) is a TOCTOU.

Fix it by open(2), fstat(2) and then fcntl(O_APPEND) if user requested
a partial content. Simplify code and reduce imsg bouncing.