Delay creating output file until url_get succeeds.
If a resume download is requested, try opening file with O_WRONLY|O_APPEND
and determine the offset required.

If the server doesn't support Range requests, truncate the file.

If above open(2) fails(file doesn't exist) or if it's a non-resume
download, then and only then create a file with O_CREAT.

http(1) will no longer create gratuitous files if HTTP response is
not 2xx.